% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INTERNAL.R
\name{.dvfisk_endpoint}
\alias{.dvfisk_endpoint}
\title{Internal function to get the base request for the SersVix API}
\usage{
.dvfisk_endpoint(what, endpoint)
}
\arguments{
\item{what}{one of SersVix and SersIndivid}

\item{endpoint}{name of the endpoint (must be defined for the API named by \code{what})}
}
\value{
A httr2 request object
}
\description{
Internal function to get the base request for the SersVix API
}
\examples{
\dontrun{
req <- .dvfiskendpoint("SersVix", "lan")
}
}
\keyword{internal}
