% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INTERNAL.R
\name{.dvfisk_get_body}
\alias{.dvfisk_get_body}
\title{Internal function that returns the body of a request}
\usage{
.dvfisk_get_body(req)
}
\arguments{
\item{req}{A httr2 request object}
}
\value{
a data.frame with the body of the request
}
\description{
Internal function that returns the body of a request
}
\examples{
\dontrun{
body <- .dvfisk_get_body(req)
}
}
\keyword{internal}
