% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{.sers_data_endpoint}
\alias{.sers_data_endpoint}
\title{Internal function to get the base request for the SERS API}
\usage{
.sers_data_endpoint(what)
}
\arguments{
\item{one}{of "lan", "kommuner", "huvudavrinningsomraden", "rapport", the defined
endpoints in the SersVix API}
}
\value{
A httr2 request object
}
\description{
Internal function to get the base request for the SERS API
}
\examples{
\dontrun{
req <- .sers_data_endpoint("lan")
}
}
\keyword{internal}
