% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SersVix.R
\name{.SersVix_endpoint}
\alias{.SersVix_endpoint}
\title{Internal function to get the base request for the SersVix API}
\usage{
.SersVix_endpoint(what)
}
\arguments{
\item{what}{one of "lan", "kommuner", "huvudavrinningsomraden", "rapport", the defined
endpoints in the SersVix API}
}
\value{
A httr2 request object
}
\description{
Internal function to get the base request for the SersVix API
}
\examples{
\dontrun{
req <- .SersVix_endpoint("lan")
}
}
\keyword{internal}
