% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvfisk-package.R
\docType{package}
\name{dvfisk-package}
\alias{dvfisk}
\alias{dvfisk-package}
\title{dvfisk: Query and Get Data from Various API:s at https://dvfisk.slu.se}
\description{
To learn more about \code{dvfisk}, start with the vignettes:
\code{browseVignettes(package = "dvfisk")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://kagervall.github.io/dvfisk/}
  \item \url{https://github.com/kagervall/dvfisk}
  \item Report bugs at \url{https://github.com/kagervall/dvfisk/issues}
}

}
\author{
\strong{Maintainer}: Anders Kagervall \email{anders.kagervall@slu.se} (\href{https://orcid.org/0000-0003-4790-2696}{ORCID})

}
\keyword{internal}
