% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPTIONS.R
\name{dvfisk_options}
\alias{dvfisk_options}
\title{Set or get options for the \code{dvfisk} package}
\usage{
dvfisk_options(...)
}
\arguments{
\item{...}{Option names to retrieve option values or \code{[key]=[value]} pairs to set options.}
}
\description{
Set or get options for the \code{dvfisk} package
}
\section{Supported options}{

Supported keys:
\itemize{
\item{\code{SersVix_url} Base URL for the API (Default: "https://dvfisk.slu.se/api/V1/sers/data-vix")}
}
}

\examples{
# Show default
dvfisk_options()

dvfisk_options(SersVix_url = "https://testing.slu.se/api/V1/sers/data-vix") # Change another server
dvfisk_options()$SersVix_url # Show the new value
settings::reset(dvfisk_options) # Reset to defaults

}
