% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KulFangst.R
\name{kul_fangst_ar}
\alias{kul_fangst_ar}
\title{Return the years when survey data available for the site specified by \code{lan}, \code{lokalId} and \code{fiske}}
\usage{
kul_fangst_ar(lan, lokalId, fiske)
}
\arguments{
\item{lan}{character, name of the county (län) where  site (lokalId) is located}

\item{lokalId}{character, id of the site (lokalId) to get years for}

\item{fiske}{character, type of survey (fiske) to get years for}
}
\value{
A vector of integers with the years when survey data available for the site (lokalId)
}
\description{
Return the years when survey data available for the site specified by \code{lan}, \code{lokalId} and \code{fiske}
}
\examples{
kul_fangst_ar("Blekinge län", 204, "Yngelinventering")

}
\seealso{
Description of
\url{https://www.slu.se/en/departments/aquatic-resources1/databases/database-for-coastal-fish-kul/},
description of API \url{https://dvfisk.slu.se/swagger/index.html}

Other KulFangst API functions: 
\code{\link{kul_fangst_fisken}()},
\code{\link{kul_fangst_lan}()},
\code{\link{kul_fangst_lokaler}()},
\code{\link{kul_fangst_rapport}()}
}
\concept{KulFangst API functions}
