% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KulFangst.R
\name{kul_fangst_fisken}
\alias{kul_fangst_fisken}
\title{Return the types of surveys available for the site specified by \code{lan} and \code{lokalId}}
\usage{
kul_fangst_fisken(lan, lokalId)
}
\arguments{
\item{lan}{character, name of the county (län) where  site (lokalId) is located}

\item{lokalId}{integer, id of the site (lokalId) to get types of surveys for}
}
\value{
A vector of characters with the types of surveys available for the site (lokalId)
}
\description{
Return the types of surveys available for the site specified by \code{lan} and \code{lokalId}
}
\examples{
# This example will access the Internet to get the data and will not work offline
kul_fangst_fisken("Blekinge län", 204)

}
\seealso{
Description of
\url{https://www.slu.se/en/departments/aquatic-resources1/databases/database-for-coastal-fish-kul/},
description of API \url{https://dvfisk.slu.se/swagger/index.html}

Other KulFangst API functions: 
\code{\link{kul_fangst_ar}()},
\code{\link{kul_fangst_lan}()},
\code{\link{kul_fangst_lokaler}()},
\code{\link{kul_fangst_rapport}()}
}
\concept{KulFangst API functions}
