% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KulFangst.R
\name{kul_fangst_lan}
\alias{kul_fangst_lan}
\title{Return counties (län) known in the KUL database}
\usage{
kul_fangst_lan()
}
\value{
A vector of type character containing names of counties (län) known in the KUL database
}
\description{
Return counties (län) known in the KUL database
}
\examples{
# This example will access the Internet to get the data and will not work offline
counties <- kul_fangst_lan()
counties
}
\seealso{
Description of
\url{https://www.slu.se/en/departments/aquatic-resources1/databases/database-for-coastal-fish-kul/},
description of API \url{https://dvfisk.slu.se/swagger/index.html}

Other KulFangst API functions: 
\code{\link{kul_fangst_ar}()},
\code{\link{kul_fangst_fisken}()},
\code{\link{kul_fangst_lokaler}()},
\code{\link{kul_fangst_rapport}()}
}
\concept{KulFangst API functions}
