% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KulFangst.R
\name{kul_fangst_lokaler}
\alias{kul_fangst_lokaler}
\title{Return sites (lokaler) in a county known in the KUL database}
\usage{
kul_fangst_lokaler(lan)
}
\arguments{
\item{lan}{character, name of the county (län) to get sites (lokaler) for}
}
\value{
A data.frame with columns \code{id} and \code{name} the sites (lokaler) in the county (län)
}
\description{
Return sites (lokaler) in a county known in the KUL database
}
\examples{
kul_fangst_lokaler("Blekinge län")
}
\seealso{
Description of
\url{https://www.slu.se/en/departments/aquatic-resources1/databases/database-for-coastal-fish-kul/},
description of API \url{https://dvfisk.slu.se/swagger/index.html}

Other KulFangst API functions: 
\code{\link{kul_fangst_ar}()},
\code{\link{kul_fangst_fisken}()},
\code{\link{kul_fangst_lan}()},
\code{\link{kul_fangst_rapport}()}
}
\concept{KulFangst API functions}
