% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KulLangd.R
\name{kul_langd_lan}
\alias{kul_langd_lan}
\title{Return counties (län) known in the KUL database}
\usage{
kul_langd_lan()
}
\value{
A vector of type character containing names of counties (län) known in the KUL database
}
\description{
Return counties (län) known in the KUL database
}
\examples{
# This example will access the Internet to get the data and will not work offline
counties <- kul_langd_lan()
counties
}
\seealso{
Description of
\url{https://www.slu.se/en/departments/aquatic-resources1/databases/database-for-coastal-fish-kul/},
description of API \url{https://dvfisk.slu.se/swagger/index.html}

Other KulLangd API functions: 
\code{\link{kul_langd_ar}()},
\code{\link{kul_langd_fisken}()},
\code{\link{kul_langd_lokaler}()},
\code{\link{kul_langd_rapport}()}
}
\concept{KulLangd API functions}
