% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorsAggregerad.R
\name{nors_aggregerad_huvudavrinningsomraden}
\alias{nors_aggregerad_huvudavrinningsomraden}
\title{Get the main drainage areas (huvudavrinningsområden) in Sweden known in the NORS database}
\usage{
nors_aggregerad_huvudavrinningsomraden()
}
\value{
A data frame with columns \code{haroNr} and \code{huvudavrinningsomrade} with the main
drainage areas (huvudavrinningsområden) in Sweden
}
\description{
This function returns the main drainage areas (huvudavrinningsområden) in Sweden.
}
\examples{
# This example will access the Internet to get the data and will not work offline
huvudavrinningsomraden <- nors_aggregerad_huvudavrinningsomraden()
head(huvudavrinningsomraden)

}
\seealso{
Description of \url{https://www.slu.se/en/departments/aquatic-resources1/databases/national-register-of-survey-test-fishing-nors/},
description of API \url{https://dvfisk.slu.se/swagger/index.html}

Other NorsAggregerad API functions: 
\code{\link{nors_aggregerad_lan}()},
\code{\link{nors_aggregerad_rapport}()}
}
\concept{NorsAggregerad API functions}
