% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorsFangst.R
\name{nors_fangst_rapport}
\alias{nors_fangst_rapport}
\title{Get aggregated data from the Nors API}
\usage{
nors_fangst_rapport(lan, kommun = NULL, vatten = NULL)
}
\arguments{
\item{lan}{character string with the name of a county (län) in Sweden}

\item{kommun}{character string with the name of a municipality in county (lan) in Sweden}

\item{vatten}{character string with the name a lake (vatten) in Sweden}
}
\value{
A data frame with survey  data. There are 13 columns in the data frame
consult NORS-documentation.
}
\description{
This function returns electrofishing data from the Nors API. You can filter the data by
county (Lan), main drainage area (HaroNr) or lake (Sjo). The function
does not check if the parameters are valid and it is possible to give combinations
that does not make sense, e.g. a \code{haroNr} and \code{Lan} that does not intersect.
It is also possible send requests that return a lot of data, so be careful.
}
\examples{
# This example will access the Internet to get the data and will not work offline
# Get all data from municipality "Hylte" in county "Hallands Län".
data <- nors_fangst_rapport(lan = "Hallands Län", kommun = "Hylte")
head(data)
dim(data)
}
\seealso{
Description of \url{https://www.slu.se/en/departments/aquatic-resources1/databases/national-register-of-survey-test-fishing-nors/},
description of API \url{https://dvfisk.slu.se/swagger/index.html}

Other NorsFangst API functions: 
\code{\link{nors_fangst_kommuner}()},
\code{\link{nors_fangst_lan}()}
}
\concept{NorsFangst API functions}
