% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorsIndivid.R
\name{nors_individ_kommuner}
\alias{nors_individ_kommuner}
\title{Return the municipalities (kommun) in Sweden known in the NORS database}
\usage{
nors_individ_kommuner(lan)
}
\arguments{
\item{lan}{character string with the name of a county (län) in Sweden}
}
\value{
A list of municipalities (kommun) in specified county (län).
}
\description{
This function returns municipalities (kommun) in  a county in Sweden. Only
correctly spelled names of counties (län) will generate a result.
}
\examples{
# This example will access the Internet to get the data and will not work offline
nors_individ_kommuner()

}
\seealso{
Description of \url{https://www.slu.se/en/departments/aquatic-resources1/databases/national-register-of-survey-test-fishing-nors/},
description of API \url{https://dvfisk.slu.se/swagger/index.html}

Other NorsIndivid API functions: 
\code{\link{nors_individ_lan}()},
\code{\link{nors_individ_rapport}()},
\code{\link{nors_individ_vatten}()}
}
\concept{NorsIndivid API functions}
