% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorsIndivid.R
\name{nors_individ_lan}
\alias{nors_individ_lan}
\title{Return counties (län) known in the NORS database}
\usage{
nors_individ_lan()
}
\value{
A vector of type character containing names of counties (län) known in the NORS database
}
\description{
Return counties (län) known in the NORS database
}
\examples{
# This example will access the Internet to get the data and will not work offline
counties <- nors_individ_lan()
counties
}
\seealso{
Description of
\url{https://www.slu.se/en/departments/aquatic-resources1/databases/national-register-of-survey-test-fishing-nors/},
description of API \url{https://dvfisk.slu.se/swagger/index.html}

Other NorsIndivid API functions: 
\code{\link{nors_individ_kommuner}()},
\code{\link{nors_individ_rapport}()},
\code{\link{nors_individ_vatten}()}
}
\concept{NorsIndivid API functions}
