% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorsIndivid.R
\name{nors_individ_rapport}
\alias{nors_individ_rapport}
\title{Get individ data from the Nors API}
\usage{
nors_individ_rapport(lan, kommun = NULL, vatten = NULL)
}
\arguments{
\item{lan}{character string with the name of a county (län) in Sweden}

\item{kommun}{character string with the name of a municipality in county (lan) in Sweden}

\item{vatten}{character string with the name a lake (vatten) in Sweden}
}
\value{
A data frame with survey  data. There are 7 columns in the data frame
consult NORS-documentation.
}
\description{
This function returns electrofishing data from the Nors API. You can filter the data by
county (Lan), main drainage area (HaroNr) or lake (Sjo). The function
does not check if the parameters are valid and it is possible to give combinations
that does not make sense, e.g. a \code{haroNr} and \code{Lan} that does not intersect.
It is also possible send requests that return a lot of data, so be careful.
}
\examples{
# This example will access the Internet to get the data and will not work offline
# Get all data from lake "626842-147473 Flaken" in county "Kalmar Län".
data <- nors_individ_rapport("Kalmar Län", vatten = "626842-147473 Flaken")
head(data)
dim(data)
}
\seealso{
Description of \url{https://www.slu.se/en/departments/aquatic-resources1/databases/national-register-of-survey-test-fishing-nors/},
description of API \url{https://dvfisk.slu.se/swagger/index.html}

Other NorsIndivid API functions: 
\code{\link{nors_individ_kommuner}()},
\code{\link{nors_individ_lan}()},
\code{\link{nors_individ_vatten}()}
}
\concept{NorsIndivid API functions}
