% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorsIndivid.R
\name{nors_individ_vatten}
\alias{nors_individ_vatten}
\title{Returns lakes (vatten) in  a county (and optionally municipality )in Sweden.}
\usage{
nors_individ_vatten(lan, kommun = NULL)
}
\arguments{
\item{lan}{character string with the name of a county (län) in Sweden}

\item{kommun}{character string with the name of a municipality in county (lan) in Sweden}
}
\value{
A list of lakes (vatten) in specified county (län).
}
\description{
Returns lakes (vatten) in  a county (and optionally municipality )in Sweden.
}
\examples{
# This example will access the Internet to get the data and will not work offline
nors_individ_vatten(lan = "Uppsala Län", kommun = "Älvkarleby")

}
\seealso{
Description of \url{https://www.slu.se/en/departments/aquatic-resources1/databases/national-register-of-survey-test-fishing-nors/},
description of API \url{https://dvfisk.slu.se/swagger/index.html}

Other NorsIndivid API functions: 
\code{\link{nors_individ_kommuner}()},
\code{\link{nors_individ_lan}()},
\code{\link{nors_individ_rapport}()}
}
\concept{NorsIndivid API functions}
