% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SersAggregerad.R
\name{sers_aggregerad_huvudavrinningsomraden}
\alias{sers_aggregerad_huvudavrinningsomraden}
\title{Get the main drainage areas (huvudavrinningsområden) in Sweden known in the SERS database}
\usage{
sers_aggregerad_huvudavrinningsomraden()
}
\value{
A data frame with columns \code{haroNr} and \code{huvudavrinningsomrade} with the main
drainage areas (huvudavrinningsområden) in Sweden
}
\description{
This function returns the main drainage areas (huvudavrinningsområden) in Sweden.
}
\examples{
# This example will access the Internet to get the data and will not work offline
huvudavrinningsomraden <- sers_aggregerad_huvudavrinningsomraden()
head(huvudavrinningsomraden)

}
\seealso{
Description of \url{https://www.slu.se/en/departments/aquatic-resources1/databases/database-for-testfishing-in-streams/},
description of API \url{https://dvfisk.slu.se/swagger/index.html}

Other SersAggregerad API functions: 
\code{\link{sers_aggregerad_lan}()},
\code{\link{sers_aggregerad_rapport}()}
}
\concept{SersAggregerad API functions}
