% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SersAggregerad.R
\name{sers_aggregerad_rapport}
\alias{sers_aggregerad_rapport}
\title{Get electrofishing data from the SERS API}
\usage{
sers_aggregerad_rapport(Lan = NULL, haroNr = NULL, Vattendrag = NULL)
}
\arguments{
\item{Lan}{character string with the name of a county (län) in Sweden}

\item{haroNr}{integer with the main drainage area (huvudavrinningsområde) number}

\item{Vattendrag}{character string with the name a river (vattendrag) in Sweden}
}
\value{
A data frame with the electrofishing data. There are 30 columns in the data frame
consult SERS-documentation.
}
\description{
This function returns electrofishing data from the SERS API. You can filter the data by
county (lan), municipality (kommun), main drainage area (haroNr). The function
does not check if the parameters are valid and it is possible to give combinations
that does not make sense, e.g. a \code{haroNr} and \code{Lan} that does not intersect.
It is also possible send requests that return a lot of data, so be careful.
}
\examples{
# This example will access the Internet to get the data and will not work offline
# Get all data from the main drainage area 24000 (Rickleån).
data <- sers_aggregerad_rapport(haroNr = 24000)
head(data)
dim(data)
}
\seealso{
Description of \url{https://www.slu.se/en/departments/aquatic-resources1/databases/database-for-testfishing-in-streams/},
description of API \url{https://dvfisk.slu.se/swagger/index.html}

Other SersAggregerad API functions: 
\code{\link{sers_aggregerad_huvudavrinningsomraden}()},
\code{\link{sers_aggregerad_lan}()}
}
\concept{SersAggregerad API functions}
