% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SersVix.R
\name{sers_data_vix_huvudavrinningsomraden}
\alias{sers_data_vix_huvudavrinningsomraden}
\title{Get the main drainage areas (huvudavrinningsområden) in Sweden known in the SERS database}
\usage{
sers_data_vix_huvudavrinningsomraden()
}
\value{
A data frame with columns \code{haroNr} and \code{huvudavrinningsomrade} with the main
drainage areas (huvudavrinningsområden) in Sweden
}
\description{
This function returns the main drainage areas (huvudavrinningsområden) in Sweden.
}
\examples{
# This example will access the Internet to get the data and will not work offline
huvudavrinningsomraden <- sers_data_vix_huvudavrinningsomraden()
head(huvudavrinningsomraden)

}
