% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SersVix.R
\name{sers_data_vix_kommuner}
\alias{sers_data_vix_kommuner}
\title{Get the municipalities (kommuner) in a county (län)}
\usage{
sers_data_vix_kommuner(lan)
}
\arguments{
\item{lan}{A character string with the name of a county (län) in Sweden}
}
\value{
A vector of type character containing names of municipalities (kommuner) in the county (län)
}
\description{
This function returns the municipalities (kommuner) in a county (län) in Sweden.
The parameter \code{lan} must be a valid county (län) in Sweden.
}
\examples{
# This example will access the Internet to get the data and will not work offline
kommuner <- sers_data_vix_kommuner("Västra Götaland")
kommuner
}
