% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{sers_data_vix_rapport}
\alias{sers_data_vix_rapport}
\title{Get electrofishing data from the SERS API}
\usage{
sers_data_vix_rapport(
  lan = NULL,
  kommun = NULL,
  haroNr = NULL,
  startdatum = NULL,
  slutdatum = NULL
)
}
\arguments{
\item{lan}{character string with the name of a county (län) in Sweden}

\item{kommun}{character string with the name of a municipality (kommun) in Sweden}

\item{haroNr}{integer with the main drainage area (huvudavrinningsområde) number}

\item{startdatum}{character string with the start date in the format "YYYY-MM-DD"}

\item{slutdatum}{character string with the end date in the format "YYYY-MM-DD"}
}
\value{
A data frame with the electrofishing data. There are 244 columns in the data frame
consult SERS-documentation.
}
\description{
This function returns electrofishing data from the SERS API. You can filter the data by
county (lan), municipality (kommun), main drainage area (haroNr), start date (startdatum)
and end date (slutdatum). You must provide at least one of theparameters. The function
does not check if the parameters are valid and it is possible to give combinations
that does not make sense, e.g. a \code{haroNr} and \code{lan} that does not intersect.
It is also possible send requests that return a lot of data, so be careful.
}
\examples{
# This example will access the Internet to get the data and will not work offline
# Get all data from the main drainage area 86000 (Mörrumsån) in 2021
data <- sers_data_vix_rapport(haroNr = 86000, startdatum = "2021-01-01", slutdatum = "2021-12-31")
head(data)
dim(data)
}
