% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SersVix.R
\name{sers_vix_lan}
\alias{sers_vix_lan}
\title{Return counties (län) known in the SERS database}
\usage{
sers_vix_lan()
}
\value{
A vector of type character containing names of counties (län) known in the SERS database
}
\description{
Return counties (län) known in the SERS database
}
\examples{
# This example will access the Internet to get the data and will not work offline
counties <- sers_vix_lan()
counties
}
\seealso{
Description of
\url{https://www.slu.se/en/departments/aquatic-resources1/databases/database-for-testfishing-in-streams/},
description of API \url{https://dvfisk.slu.se/swagger/index.html}

Other SersVix API functions: 
\code{\link{sers_vix_huvudavrinningsomraden}()},
\code{\link{sers_vix_kommuner}()},
\code{\link{sers_vix_rapport}()}
}
\concept{SersVix API functions}
