% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPTIONS.R
\name{sersapi_options}
\alias{sersapi_options}
\title{Set or get options for the \code{sersapi} package}
\usage{
sersapi_options(...)
}
\arguments{
\item{...}{Option names to retrieve option values or \code{[key]=[value]} pairs to set options.}
}
\description{
Set or get options for the \code{sersapi} package
}
\section{Supported options}{

Supported keys:
\itemize{
\item{\code{base_url} Base URL for the API (Default: "https://dvfisk.slu.se/api/V1/sers/data-vix")}
}
}

\examples{
# Show default
sersapi_options()

sersapi_options(base_url = "https://testing.slu.se/api/V1/sers/data-vix") # Change another server
sersapi_options()$base_url # Show the new value
settings::reset(sersapi_options) # Reset to defaults

}
