% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SITES_documentation.R
\docType{data}
\name{sites}
\alias{sites}
\alias{sites_sf}
\title{dvfisk::sites A data frame with information about electrofishing sites in SERS}
\format{
Columns in \code{dvfisk::sites} are:
\describe{
\item{lan}{Länen kodas 1 till 25 (Stockholms till Norrbottens län).}
\item{hflodomr}{Huvudflodområdesnummer enligt SMHI:s numrering (Vattendragsregistret).}
\item{biflnr}{Biflödesnummer enligt SMHI (Vattendragsregistret).}
\item{vdragnam}{Vattendragets namn.}
\item{xkoorvdr}{Vattendragets mynningskoordinat i X-led (nord-syd) enligt SMHI (Vattendragsregistret).}
\item{ykoorvdr}{Vattendragets mynningskoordinat i Y-led (väst-öst) enligt SMHI (Vattendragsregistret).}
\item{lokalnam}{Elfiskelokalens namn}
\item{lokalnr}{Elfiskelokalens nummer (om blankt saknas lokalnummer för aktuell elfiskelokal).}
\item{xkoorlok}{X-koordinat (nord-syd; sexsiffrig RT90, 2.5 gon väst) för elfiskelokalens nedersta punkt.}
\item{ykoorlok}{Y-koordinat (väst-öst; sexsiffrig RT90, 2.5 gon väst) för elfiskelokalens nedersta punkt.}
\item{hoh}{Höjd över havet i meter.}
\item{avriomrk}{Avrinningsområdets storlek uppströms elfiskelokalen (km2). Anges enbart i fyra klasser  (<10, <100, <1000 resp >1000).}
\item{andsjopr}{Andelen sjöar i \% av avrinningsområdet uppströms elfiskelokalen. Anges enbart i fyra klasser (<1, <5, <10, >10).}
\item{avstupp}{Avstånd (km) till uppströms sjö eller större sel (lugnparti) om minst 1 hektar i huvudfåran.}
\item{avstner}{Avstånd (km) till nedströms sjö eller större sel (lugnparti) om minst 1 hektar i huvudfåran.}
\item{kommunnr}{Anger vilken kommun som elfiskelokalen ligger i enligt SCB.}
\item{kommun}{Namn på kommunen där elfiskelokalen är lokaliserad.}
\item{topokart}{Här anges vilken topografisk karta elfiskelokalen ligger på, ex 10B SO. Lämnas tomt om ej angivet.}
\item{x_rt90_sju}{X-koordinat (nord-syd; sjusiffrig RT90, 2.5 gon väst) för elfiskelokalens nedersta punkt.}
\item{y_rt90_sju}{Y-koordinat (väst-öst; sjusiffrig RT90, 2.5 gon väst) för elfiskelokalens nedersta punkt.}
\item{s99tm_n}{Koordinat (nord-syd, Sweref99 TM) för elfiskelokalens nedersta punkt. Obs. ska ej anges på elfiskeprotokoll!}
\item{s99tm_e}{Koordinat (väst-öst, Sweref99 TM) för elfiskelokalens nedersta punkt. Obs. ska ej anges på elfiskeprotokoll!}
\item{ddlat}{Latitud för elfiskelokalen (WGS84) för elfiskelokalens nedersta punkt. Obs. ska ej anges på elfiskeprotokoll!}
\item{ddlong}{Latitud för elfiskelokalen (WGS84) för elfiskelokalens nedersta punkt. Obs. ska ej anges på elfiskeprotokoll!}
}

An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 21749 rows and 3 columns.
}
\source{
Swedish Electrofishing Register (SERS)
\url{https://www.slu.se/en/departments/aquatic-resources1/databases/database-for-testfishing-in-streams/}
}
\usage{
sites

sites_sf
}
\description{
Data from the Swedish Electrofishing Register (SERS) containing information about electrofishing sites.
The documentation is based on the Excel file provided by SLU and avaliable in Swedish only.

Data frame \code{dvfisk::sites} is a list of sites defined in SERS  with 23
variables described below.

\code{dvfisk::sites_sf} is a simple feature points object for the sites in \code{dvfisk::sites}
use \code{dplyr::left_join(sites, sites_sf, by = dplyr::join_by(XKOORLOK, YKOORLOK))}
to join the data frames. You might want to use \code{library(sf)} in your script to
work with the simple feature object.
}
\concept{Datasets}
\keyword{datasets}
