% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorsFangst.R
\name{nors_fangst_kommuner}
\alias{nors_fangst_kommuner}
\title{Get the municipalities (kommun) in Sweden known in the NORS database}
\usage{
nors_fangst_kommuner(lan)
}
\arguments{
\item{lan}{character string with the name of a county (län) in Sweden}
}
\value{
A list of municipalities (kommun) in specified county (län).
}
\description{
This function returns municipalities (kommun) in  a county in Sweden.
}
\examples{
# This example will access the Internet to get the data and will not work offline
nors_fangst_huvudavrinningsomraden()

}
\seealso{
Description of \url{https://www.slu.se/en/departments/aquatic-resources1/databases/national-register-of-survey-test-fishing-nors/},
description of API \url{https://dvfisk.slu.se/swagger/index.html}

Other NorsFangst API functions: 
\code{\link{nors_fangst_lan}()},
\code{\link{nors_fangst_rapport}()}
}
\concept{NorsFangst API functions}
